+++ lib/server/server-internal.ts
@@ -1,7 +1,19 @@
 import env from "./environment";
 import { InstantBanditServer, InstantBanditServerOptions } from "./server-types";
 import { buildInstantBanditServer } from "./server";
+import { getJsonSiteBackend } from "./backends/json-sites";
+import { getRedisBackend } from "./backends/redis";
 
+
+const redis = getRedisBackend();
+const json = getJsonSiteBackend();
+
+const DEFAULT_DEV_SERVER_OPTIONS: Partial<InstantBanditServerOptions> = {
+    clientOrigins: (env.IB_ORIGINS_ALLOWLIST ?? env.IB_BASE_API_URL),
+    sessions: redis,
+    metrics: redis,
+    models: json,
+};
 
 /**
  * This server helper intended is intended for use by the Instant Bandit repo.
@@ -13,7 +25,7 @@ import { buildInstantBanditServer } from "./server";
  * @private
  */
 let internalBanditServer: InstantBanditServer;
-export function getInternalDevServer(options?: InstantBanditServerOptions) {
+export function getInternalDevServer(options: Partial<InstantBanditServerOptions> = DEFAULT_DEV_SERVER_OPTIONS) {
   if (env.isDev()) {
     if (global["internalBanditServer"]) {
       internalBanditServer = global["internalBanditServer"];
